Convert FeedProcessor to an interface and allow for Guice-injected replacement.
