+++ java/gadgets/src/main/java/org/apache/shindig/gadgets/FeedProcessor.java
@@ -39,10 +39,14 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import com.google.inject.ImplementedBy;
+
+
 /**
  * Processes RSS & Atom Feeds and converts them into JSON output.
  */
-public class FeedProcessor {
+@ImplementedBy(FeedProcessorImpl.class)
+public interface FeedProcessor {
 
   /**
    * Converts feed XML to JSON.
@@ -57,152 +61,6 @@ public class FeedProcessor {
    *            Number of entries to return.
    * @return The JSON representation of the feed.
    */
-  @SuppressWarnings("unchecked")
-  public JSONObject process(String feedUrl, String feedXml, boolean getSummaries, int numEntries)
-          throws GadgetException {
-    try {
-      SyndFeed feed = new SyndFeedInput().build(new StringReader(feedXml));
-      JSONObject json = new JSONObject();
-      json.put("Title", feed.getTitle());
-      json.put("URL", feedUrl);
-      json.put("Description", feed.getDescription());
-      json.put("Link", feed.getLink());
-
-    //Retrieve the feed image if it is available as well as an image url if the image is available.
-    if(feed.getImage() != null && !feed.getImage().getUrl().isEmpty()){
-        SyndImage feedImage = (SyndImage)feed.getImage();
-        JSONObject jsonImage = new JSONObject();
-        jsonImage.put("Url", feedImage.getUrl());
-        if(feedImage.getTitle() != null 
-                && !feedImage.getTitle().isEmpty()){
-            jsonImage.put("Title", feedImage.getTitle());
-        }
-        if(feedImage.getDescription() != null &&
-                !feedImage.getDescription().isEmpty()){
-            jsonImage.put("Description", feedImage.getDescription());
-        }
-        if(feedImage.getLink() != null && 
-        		  !feedImage.getLink().isEmpty()){
-        jsonImage.put("Link", feedImage.getLink());
-        }
-        json.put("Image", jsonImage);
-    }
-
-      
-      List<SyndPerson> authors = feed.getAuthors();
-      String jsonAuthor = null;
-      if (authors != null && !authors.isEmpty()) {
-        SyndPerson author = authors.get(0);
-        if (author.getName() != null) {
-          jsonAuthor = author.getName();
-        } else if (author.getEmail() != null) {
-          jsonAuthor = author.getEmail();
-        }
-      }
-      JSONArray entries = new JSONArray();
-      json.put("Entry", entries);
-
-      int entryCnt = 0;
-      for (Object obj : feed.getEntries()) {
-        SyndEntry e = (SyndEntry) obj;
-        if (entryCnt >= numEntries) {
-          break;
-        }
-        entryCnt++;
-
-        JSONObject entry = new JSONObject();
-        entry.put("Title", e.getTitle());
-        entry.put("Link", e.getLink());
-        if (getSummaries) {
-          if (e.getContents() != null && !e.getContents().isEmpty()) {
-            entry.put("Summary", ((SyndContent) e.getContents().get(0)).getValue());
-          } else {
-            entry.put("Summary", e.getDescription() != null ? e.getDescription().getValue() : "");
-          }
-        }
-
-        if (e.getUpdatedDate() != null) {
-          entry.put("Date", e.getUpdatedDate().getTime());
-        } else if (e.getPublishedDate() != null) {
-          entry.put("Date", e.getPublishedDate().getTime());
-        } else {
-          entry.put("Date", 0);
-        }
-
-        // if no author at feed level, use the first entry author
-        if (jsonAuthor == null) {
-          jsonAuthor = e.getAuthor();
-        }
-
-        JSONObject media = new JSONObject();
-        MediaEntryModule mediaModule = (MediaEntryModule) e.getModule(MediaModule.URI);
-        if (mediaModule != null) {
-          if (mediaModule.getMediaContents().length > 0) {
-            JSONArray contents = new JSONArray();
-
-            for (MediaContent c : mediaModule.getMediaContents()) {
-              JSONObject content = new JSONObject();
-
-              if (c.getReference() instanceof UrlReference) {
-                content.put("URL", ((UrlReference) c.getReference()).getUrl().toString());
-              }
-
-              if (c.getType() != null) {
-                content.put("Type", c.getType());
-              }
-
-              if (c.getWidth() != null) {
-                content.put("Width", c.getWidth());
-              }
-
-              if (c.getHeight() != null) {
-                content.put("Height", c.getHeight());
-              }
-
-              contents.put(content);
-            }
-
-            media.put("Contents", contents);
-          }
-
-          if (mediaModule.getMetadata() != null) {
-            if (mediaModule.getMetadata().getThumbnail().length > 0) {
-              // "If multiple thumbnails are included, it is assumed that they are in order of importance"
-              // Only use the first thumbnail for simplicity's
-              // sake
-
-              JSONObject thumbnail = new JSONObject();
-
-              Thumbnail t = mediaModule.getMetadata().getThumbnail()[0];
-              thumbnail.put("URL", t.getUrl().toString());
-
-              if (t.getWidth() != null) {
-                thumbnail.put("Width", t.getWidth());
-              }
-
-              if (t.getHeight() != null) {
-                thumbnail.put("Height", t.getHeight());
-              }
-
-              media.put("Thumbnail", thumbnail);
-            }
-          }
-        }
-
-        entry.put("Media", media);
-
-        entries.put(entry);
-      }
-
-      json.put("Author", (jsonAuthor != null) ? jsonAuthor : "");
-      return json;
-    } catch (JSONException e) {
-      // This shouldn't ever happen.
-      throw new RuntimeException(e);
-    } catch (FeedException e) {
-      throw new GadgetException(GadgetException.Code.MALFORMED_XML_DOCUMENT, e, HttpResponse.SC_BAD_GATEWAY);
-    } catch (IllegalArgumentException e) {
-      throw new GadgetException(GadgetException.Code.MALFORMED_XML_DOCUMENT, e, HttpResponse.SC_BAD_GATEWAY);
-    }
-  }
+  JSONObject process(String feedUrl, String feedXml, boolean getSummaries, int numEntries)
+          throws GadgetException;
 }
