+++ java/gadgets/src/test/java/org/apache/shindig/gadgets/servlet/ServletTestFixture.java
@@ -20,6 +20,8 @@ package org.apache.shindig.gadgets.servlet;
 
 import org.apache.shindig.common.EasyMockTestCase;
 import org.apache.shindig.common.servlet.HttpServletResponseRecorder;
+import org.apache.shindig.gadgets.FeedProcessor;
+import org.apache.shindig.gadgets.FeedProcessorImpl;
 import org.apache.shindig.gadgets.LockedDomainService;
 import org.apache.shindig.gadgets.http.RequestPipeline;
 import org.apache.shindig.gadgets.rewrite.CaptureRewriter;
@@ -31,6 +33,8 @@ import java.util.Arrays;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import com.google.inject.Provider;
 
 /**
  * Contains everything needed for making servlet requests, plus a bunch of stuff that shouldn't be
@@ -47,4 +51,9 @@ public abstract class ServletTestFixture extends EasyMockTestCase {
   public final HttpServletResponse response = mock(HttpServletResponse.class);
   public final HttpServletResponseRecorder recorder = new HttpServletResponseRecorder(response);
   public final LockedDomainService lockedDomainService = mock(LockedDomainService.class);
+  public final Provider<FeedProcessor> feedProcessorProvider = new Provider<FeedProcessor>() {
+        public FeedProcessor get() {
+            return new FeedProcessorImpl();
+        }
+  };
 }
